+++ scripts/dev-data/no-projects.config.ts
@@ -13,7 +13,7 @@ export default defineConfig({
   reporter: [
     ['dot'],
     [
-      `${projectRoot}/src/reporter.ts`,
+      `${projectRoot}/src/index.ts`,
       {
         outputFile: `${projectRoot}/src/report/with-no-projects.tpl.html`,
       },