+++ test/e2e/playwright.config.ts
@@ -1,7 +1,6 @@
 import { defineConfig } from '@playwright/test';
-import { timelineReporter } from '../../dist/index.js';
 
 export default defineConfig({
   testDir: './specs',
-  reporter: [timelineReporter({ outputFile: './timeline-report/index.html' })],
+  reporter: [['../../dist/index.js', { outputFile: './timeline-report/index.html' }]],
 });